+++ src/main/java/com/mineplus/MineplusPlugin.java
@@ -42,7 +42,6 @@ public final class MineplusPlugin extends JavaPlugin {
         virtualBlockManager = new VirtualBlockManager();
         virtualBlockManager.updateSettings(configManager.getConfig().getVirtualRendering());
         virtualBlockManager.updateTexelSettings(configManager.getConfig().getTexelBaking());
-        virtualBlockManager.updateVoxelSettings(configManager.getConfig().getVoxelRendering());
         attachDisplayTransport();
         virtualBlockManager.loadModels(this);
 
@@ -142,7 +141,6 @@ public final class MineplusPlugin extends JavaPlugin {
         if (virtualBlockManager != null && configManager != null) {
             virtualBlockManager.updateSettings(configManager.getConfig().getVirtualRendering());
             virtualBlockManager.updateTexelSettings(configManager.getConfig().getTexelBaking());
-            virtualBlockManager.updateVoxelSettings(configManager.getConfig().getVoxelRendering());
         }
         if (context != null && configManager != null) {
             context.infrastructureEngine().updateAnimationSettings(configManager.getConfig().getAnimation());